Removed gateway Linking
