+++ test/gateway/gateway_base/activate_gateway.js
@@ -13,12 +13,10 @@ contract('GatewayBase.sol', function (accounts) {
     beforeEach(async function () {
 
       let core = accounts[0]
-        , messageBus = accounts[1]
         , bounty = new BN(100);
 
       gatewayBaseInstance = await GatewayBase.new(
         core,
-        messageBus,
         bounty,
         organisation
       );
@@ -49,12 +47,10 @@ contract('GatewayBase.sol', function (accounts) {
 
     beforeEach(async function () {
       let core = accounts[0]
-        , messageBus = accounts[1]
         , bounty = new BN(100);
 
       gatewayBaseInstance = await GatewayBase.new(
         core,
-        messageBus,
         bounty,
         organisation
       );
